# public/apidocs/api-docs.json
{  
  "swagger":"2.0",
  "info":{  
    "title":"Cents API",
    "description":"Examine your life decisions RESTfully.",
    "version":"2.0.0"
  },
  "paths":{  
...
    "/users/{id}/completed":{  
      "get":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"The user was found."
          },
          "404":{  
            "description":"No such user was found."
          }
        },
        "parameters":[  
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          }, 
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          }
        ],
        "description":"",
        "summary":"Shows which sections a user has completed.",
        "operationId":"Api::V2::UsersController#show_completed"
      },
      "post":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"User was found."
          },
          "404":{  
            "description":"No such user was found."
          }
        },
        "description":"",
        "summary":"Record that a user has completed a section.",
        "operationId":"Api::V2::UserController#create_completed",
        "parameters":[
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },   
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"body",
            "description":"object with section field",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/Completed_body"
            }
          }
        ]
      }
    },
    "/users/{id}/query":{  
      "get":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"The user was found."
          },
          "404":{  
            "description":"No such user was found."
          }
        },
        "parameters":[ 
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },  
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          }
        ],
        "description":"",
        "summary":"Shows a user's past queries.",
        "operationId":"Api::V2::UsersController#show_query"
      },
      "post":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"User was found."
          },
          "404":{  
            "description":"No such user was found."
          }
        },
        "description":"",
        "summary":"Record that a user has made a query.",
        "operationId":"Api::V2::UserController#create_query",
        "parameters":[
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },   
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"body",
            "description":"object with query field",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/QueryBody"
            }
          }
        ]
      }
    },
    "/users/{id}/ratings":{  
      "get":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"The user was found."
          },
          "404":{  
            "description":"No such user was found."
          }
        },
        "parameters":[
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },   
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          }
        ],
        "description":"",
        "summary":"Shows all items a user has rated.",
        "operationId":"Api::V2::UsersController#show_ratings"
      }
    },
...
    "/users/{id}/spending_breakdown/{category}":{  
      "patch":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"Data was deleted."
          },
          "400":{  
            "description":"Each field object must have a name and value."
          },
          "404":{  
            "description":"No such user was found."
          },
          "500":{  
            "description":"Server error while saving."
          }
        },
        "description":"Use this route to save a few individual spending breakdown items.",
        "summary":"Saves new spending breakdown data.",
        "operationId":"Api::V2::SpendingBreakdownController#update",
        "parameters":[
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },   
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"path",
            "description":"a spending breakdown category",
            "name":"category",
            "required":true,
            "type":"string",
            "enum":[  
              "default",
              "student",
              "custom"
            ]
          },
          {  
            "in":"body",
            "description":"object with array of spending breakdown data",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/SB_body"
            }
          }
        ]
      },
      "put":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"Data was deleted."
          },
          "400":{  
            "description":"Each field object must have a name and value."
          },
          "404":{  
            "description":"No such user was found."
          },
          "500":{  
            "description":"Server error while saving."
          }
        },
        "description":"Use this route to replace all spending breakdown data for a category with whatever you send. This route could be used to reset to a default.",
        "summary":"Updates all spending breakdown data by category.",
        "operationId":"Api::V2::SpendingBreakdownController#update_all",
        "parameters":[ 
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },  
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"path",
            "description":"a spending breakdown category",
            "name":"category",
            "required":true,
            "type":"string",
            "enum":[  
              "default",
              "student",
              "custom"
            ]
          },
          {  
            "in":"body",
            "description":"object with array of spending breakdown data",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/SB_body"
            }
          }
        ]
      },
      "get":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "404":{  
            "description":"No records were found."
          }
        },
        "description":"",
        "summary":"Gets spending breakdown data by category.",
        "operationId":"Api::V2::SpendingBreakdownController#show_category",
        "parameters":[  
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          }, 
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"path",
            "description":"a spending breakdown category",
            "name":"category",
            "required":true,
            "type":"string",
            "enum":[  
              "default",
              "student",
              "custom"
            ]
          }
        ]
      }
    },
    "/users/{id}/spending_breakdown/{category}/{name}":{  
      "delete":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"Data was deleted."
          }
        },
        "description":"",
        "summary":"Deletes a single spending breakdown item.",
        "operationId":"Api::V2::SpendingBreakdownController#destroy",
        "parameters":[
          {  
            "in":"query",
            "description":"the API key you've been assigned",
            "name":"api_key",
            "required":true,
            "type":"string"
          },   
          {  
            "in":"path",
            "description":"a user ID",
            "name":"id",
            "required":true,
            "type":"integer"
          },
          {  
            "in":"path",
            "description":"a spending breakdown category",
            "name":"category",
            "required":true,
            "type":"string",
            "enum":[  
              "default",
              "student",
              "custom"
            ]
          },
          {  
            "in":"path",
            "description":"the name of the spending breakdown item",
            "name":"name",
            "required":true,
            "type":"string"
          }
        ]
      }
    },
...
    "/users/validate":{  
      "post":{  
        "tags":[  
          "users"
        ],
        "responses":{  
          "200":{  
            "description":"Authentication successful.",
            "schema":{  
              "$ref":"#/definitions/IDReponse"
            }
          },
          "400":{  
            "description":"Authentication failed."
          }
        },
        "description":"This route will return a user ID that you can use with other user routes.",
        "summary":"Validate a username and password.",
        "operationId":"Api::V2::UsersController#validate",
        "parameters":[  
          {  
            "in":"body",
            "description":"object with password field",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/Validate_body"
            }
          }
        ]
      }
    }
  },
  "host":"trycents.com",
  "basePath":"/api/v2",
  "schemes":[  
    "https"
  ],
  "definitions":{  
...
    "Validate_body":{  
      "required":[  
        "email",
        "password"
      ],
      "properties":{  
        "email":{  
          "type":"string"
        },
        "password":{  
          "type":"string"
        }
      }
    },
...
    "SB_body":{  
      "type":"object",
      "properties":{  
        "fields":{  
          "type":"array",
          "items":{  
            "$ref":"#/definitions/Field"
          }
        }
      },
      "required":[  
        "fields"
      ]
    },
...
    "Completed_body":{  
      "type":"object",
      "properties":{  
        "section":{  
          "type":"string"
        }
      },
      "required":[  
        "section"
      ]
    },
...
    "IDReponse":{  
      "type":"object",
      "required":[  
        "id"
      ],
      "properties":{  
        "id":{  
          "type":"integer"
        }
      }
    },
...
    "QueryBody":{  
      "type":"object",
      "required":[  
        "url"
      ],
      "properties":{  
        "url":{  
          "type":"string"
        }
      }
    },
...
    "Field":{  
      "required":[  
        "name", "value"
      ],
      "properties":{  
        "name":{  
          "type":"string"
        },
        "value":{  
          "type":"string"
        }
      }
    }
  }
}